Removed pretranspiling of gulp files
Can only work after install and the presence of the presets and
babel-plugins in the node_modules dir. Cannot work with unit testing.

Yeoman reads the directory it is run in, not the source directory,
so either the .babelrc file is present or an option is passed to
babel() and then the generators complain that some packages are missing,
or no .babelrc/option, and then no transpilation anyway.