Config is now initialized by Base generator-wupjs
Used to be in ctor. So was initialized when process was launched.
But when process change dir, then Config was not set from running
directory. Therefore it may have the wrong .yo-rc.json-stable-stringify

THis was the case when running the unit tests, because context is
on each test by Mocha. But starting process is within generator-wupjs
directory, where there is already a .yo-rc.json file.

Now most tests are failing because 'Argument description is missing
or undefined' (it was picked out of the generator config file).