Check on deprecation and "peer" versions
When a package is deprecated, remove it.
If the user reinstall it, he will be warned of the deprecation.
all addDep method take 2nd argument. THis allows to generalize
filtering on canonic "^0.0.0" semver and others
Canonic semver are always upgraded. The other versions are
kept as they are.
Fixed bug of early return in _filterDeps
Fixed bad unrelated condition in _isUserDep